feature/WEBUMENIA-1440:  rename trait
